# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ src/Background/Tree.h
@@ -2,7 +2,8 @@
 #define THE_CORP_GAME_TREE_H
 
 
-#include <System/Enum.h>
+#include "..\System\Enum.h"
+#include "..\Basic\Entity.h"
 #include <SFML/Window.hpp>
 
 class Tree : public Entity {
@@ -10,7 +11,7 @@ public:
     const static int width = 267;
     const static int height = 517;
 
-    explicit Tree(sf::Vector2f coordinates, Entities type, sf::Vector2f size = {width, height});
+    explicit Tree(sf::Vector2f coordinates, Entities type, sf::Vector2f size = {(float) width, (float) height});
 };
